+++ truck-rendimpl/examples/textured-cube.rs
@@ -41,7 +41,12 @@ impl MyApp {
 }
 
 impl App for MyApp {
-    fn init(handler: &DeviceHandler) -> MyApp {
+    fn init(handler: &DeviceHandler, info: AdapterInfo) -> MyApp {
+        let sample_count = match info.backend {
+            Backend::Vulkan => 2,
+            Backend::Dx12 => 2,
+            _ => 1,
+        };
         let desc = SceneDescriptor {
             camera: MyApp::create_camera(),
             lights: vec![Light {
@@ -49,6 +54,7 @@ impl App for MyApp {
                 color: Vector3::new(1.0, 1.0, 1.0),
                 light_type: LightType::Point,
             }],
+            sample_count,
             ..Default::default()
         };
         let mut scene = Scene::new(handler.clone(), &desc);
@@ -121,6 +127,9 @@ impl App for MyApp {
             if let Some(ref prev_position) = self.prev_cursor {
                 let matrix = &mut self.scene.descriptor_mut().camera.matrix;
                 let dir2d = &position - prev_position;
+                if dir2d.so_small() {
+                    return Self::default_control_flow();
+                }
                 let mut axis = dir2d[1] * matrix[0].truncate();
                 axis += dir2d[0] * matrix[1].truncate();
                 axis /= axis.magnitude();
